Version 24
This is the first ready-for-release version, as it contains demographics collection before the practice trials and actual blocks.  This code contains states for consent and debrief forms, but they are never accessed at this point (as these steps are currently handled by @collyeeliz).

In this commit we:
  - [edit] Shifted the <kbd>Next</kbd> button down;
  - [edit] Made content text smaller (changed to 12pt);
  - [feature] Added demographics state and recorded them as metrics (including having error messages for invalid demographics).

Still to do:
  - Make the next button better (more dynamic);
  - ITI metric?
  - Code cleanup (old comments, print statements, class &Implies; ID in HTML, etc.).