WIP: begin work switching T1 and T2 logic around
We want T2 to always be a flower, and T1 to be one of face, watch, and
pareidolia.  As such, we need to combine all of these types of images
into a single T1 image array struct that already exists (however instead
of keeping track of URLs only, we will want to keep track of their names
and types as well; hence the ImageStruct object).