WIP: Continued refactor/implementation: separate types/configs
Massive diff in this commit, but a lot of the work here was spent
putting global variables into classes or interfaces, which are now
defined in types.ts and configured in config.ts (which also takes
configuration away from utils.ts so that it is all in (mostly) one
place).

I also continued work on making image arrays image structs rather
than just URLs, as this is a cleaner way to store information so that
it is extensible in the future.