WIP: Construct image lists as ImageStruct[], from which we extract URLs
Previously we were keeping track of many kinds of image lists globally,
such as numbered lists, lists of image names, and lists of their URLs.
Instead of this, I want to only keep track of a list of ImageStructs,
which contains all of the necessary information.  Small refactor was
required, including changing the preloading stimuli template.

In the interest of making it easier to change T1/T2 types (as we are
required to for this variation of the experiment), I have also made
two fields (t1Images, t2Images) in the global experiment state variable.
These will now be easily accessed globally, despite blocks.

I would like to do away with blocks entirely in this experiment, but
it may be more effor than it's worth.  We will see how the experiment
is working functionally before making such a decision.