WIP: try to fix GlobalExperimentState is not a constructor error
Upon running this code after the big refactor, it throws an error
shortly after our imports, where we construct a new
GlobalExperimentState, stating that this class (GlobalExperimentState)
is not a constructor.

After a long time debugging, I found two facts:
  1. The error is because the GlobalExperimentState is failing to
  import, and so we try to call new on an undefined value;
  2. The constructor function was not a public function.

After fixing the latter, I have tried various different methods of
importing the GlobalExperimentState class so that it may work, however
none have been successful.  I am pushing this version now as a clean
state in which we have a new state.ts file in which our
GlobalExperimentState class is stored (and can be exported as default
if needed).  If we figure out the issue and it doesn't need to be
exported as default, it might be worth putting it back into the
types.ts file.